Move weighted_average from diagnostics_utils to vcm
