+++ external/vcm/vcm/calc/__init__.py
@@ -1,4 +1,4 @@
-from .calc import apparent_heating, apparent_source, local_time
+from .calc import apparent_heating, apparent_source, local_time, weighted_average
 from .metrics import r2_score
 from .q_terms import compute_Q_terms
 from ._zenith_angle import cos_zenith_angle
@@ -12,4 +12,5 @@ __all__ = [
     "compute_Q_terms",
     "local_time",
     "cos_zenith_angle",
+    "weighted_average",
 ]
